+++ frontend/src/components/Recipe/index.ts
@@ -6,5 +6,5 @@ export { default as RecipeList } from './RecipeList/RecipeList';
 export { default as RecipeRegisterForm } from './RecipeRegisterForm/RecipeRegisterForm';
 export { default as RecipeFavorite } from './RecipeFavorite/RecipeFavorite';
 export { default as CommentItem } from './CommentItem/CommentItem';
-export { default as CommentInput } from './CommentInput/CommentInput';
+export { default as CommentForm } from './CommentForm/CommentForm';
 export { default as CommentList } from './CommentList/CommentList';
